feat: Add filter functionality to MahasiswaController@index
This commit adds filter functionality to the `index` method of `MahasiswaController`. The filters include `thn_masuk`, `thn_lulus`, and `jenjang`. The `thn_masuk` and `thn_lulus` filters allow the user to filter the results based on the year of admission and graduation, respectively. The `jenjang` filter allows the user to filter the results based on the educational level. This enhancement improves the usability of the Mahasiswa listing page.